requirements.txt
Django==4.2.3
mysqlclient==2.1.1
pytz==2023.3
sqlparse==0.4.4
gunicorn
whitenoise
dj-database-url
psycopg2-binary
django
gunicorn
python-decouple
